Export _IO_stdin_used symbol in NetworkManager.ver
This symbol is required to decide which version of certain IO functions
to use on various architectures. So we can't strip away the symbol from
the NetworkManager binary as this will lead to segfaults on those
architectures.

Thanks James Cowgill and Aurelien Jarno for the bug report and detailed
analysis.

Closes: #835550