Split introspection data for libnm into gir1.2-nm-1.0
Keep the introspection data for the deprecated libnm-glib and libnm-util
library in gir1.2-networkmanager-1.0 but split out the introspection
data for libnm into a new gir1.2-nm-1.0 package.
This will make a transition from libnm-glib/libnm-util to libnm easier.
To not break existing packages, add a dependency on gir1.2-nm-1.0 to
gir1.2-networkmanager-1.0.