Register empty NMClient and NetworkManager when loading libnm via GIR
This prevents the real modules from being loaded because they would in
turn load libnm-glib and abort() and crash as loading libnm and
libnm-glib in the same process is not allowed.

In particular this prevents the GNOME shell from crashing with
libnm-glib abort and allows gracefully disabling the extensions which
use the obsolete library.

Patch cherry-picked from upstream Git.